Changed maxsound in TreeCell from FLOAT to float

+++ src/Headers/CommunicationHandler.h
@@ -397,7 +397,7 @@ public:
 
     c.m = 0;
     c.hmax = 0;
-    c.maxsound = 0;
+    c.maxsound = 0.0f;
     c.amin = big_number;
     for (int k=0; k<ndim; k++) {
       c.bb.min[k] = big_number;
@@ -425,7 +425,7 @@ public:
       }
       c.m += p.m;
       c.hmax = max(c.hmax,p.h);
-      c.maxsound = max(c.maxsound, (FLOAT) p.sound);
+      c.maxsound = max(c.maxsound, p.sound);
       c.amin = min(c.amin,
                    sqrt(DotProduct(partdata[i].atree,partdata[i].atree,ndim)));
     }
